mysite/requirements.txt
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
Django==2.2.5
django-ckeditor==5.4.0
django-js-asset==2.0.0
Pillow==9.2.0
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
wincertstore==0.2
